feat: added node_config structs

+++ node/include/config.h
@@ -48,9 +48,9 @@ const uint8_t MAX_ID_ATTEMPTS    = 0;
 
 //// ESP specific
 RTC_DATA_ATTR bool needs_initialisation = true;
-RTC_DATA_ATTR int  boot_count           = 0;
-RTC_DATA_ATTR int  node_id              = 0;
-RTC_DATA_ATTR int  MAX_TX_RETRIES       = 5;
-RTC_DATA_ATTR int  BUFFERING_THRESHOLD  = 4;
-RTC_DATA_ATTR int  buffering_counter    = 0;
+RTC_DATA_ATTR uint8_t  boot_count           = 0;
+RTC_DATA_ATTR uint8_t  node_id              = 0;
+RTC_DATA_ATTR uint8_t  MAX_TX_RETRIES       = 5;
+RTC_DATA_ATTR uint8_t  BUFFERING_THRESHOLD  = 4;
+RTC_DATA_ATTR uint8_t  buffering_counter    = 0;
 #endif
